backend/requirements.txt
Flask>=3.0.2
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
